+++ src/main/java/com/gotcha/server/member/controller/MemberController.java
@@ -2,7 +2,7 @@ package com.gotcha.server.member.controller;
 
 import com.gotcha.server.member.dto.response.TodayInterviewResponse;
 import com.gotcha.server.auth.dto.response.RefreshTokenResponse;
-import com.gotcha.server.auth.security.MemberDetails;
+import com.gotcha.server.auth.dto.request.MemberDetails;
 import com.gotcha.server.member.dto.response.LoginResponse;
 import com.gotcha.server.member.dto.request.RefreshTokenRequest;
 import com.gotcha.server.member.dto.response.UserResponse;
